+++ lib/gql/resolvers/base-subscription-resolver.mixin.d.ts
@@ -5,7 +5,7 @@ import { RawPayload } from './create-node-notification';
 import { BaseReadableResolverInterface } from './base-resolver.interface';
 export declare function withSubscriptionResolver<TBase extends ClassType<BaseReadableResolverInterface<TApi, any, any>>, TApi extends Node<TApi>, TNotification extends NodeNotification<TApi>, TSubscriptionFilter extends SubscriptionWatchFilter>(capitalizedName: string, Base: TBase, notificationClsType: ClassType<TNotification>, subscriptionFilterClsType: ClassType<TSubscriptionFilter>, defaultScope: Scopes | Scopes[]): {
     new (...args: any[]): {
-        onChange(rawPayload: RawPayload, args: SubscriptionWatchFilter): Promise<NodeNotification<TApi>>;
+        onChange(rawPayload: RawPayload | undefined, args: SubscriptionWatchFilter): Promise<NodeNotification<TApi> | null>;
         ctx: import("../..").RumbleshipContext;
         getAll(filterBy: any): Promise<any>;
         getOne(id: string): Promise<TApi>;